remove version specific versions of the signListener as 1.18 is gone now. No longer create SignTranslateEvent but still react to it
