+++ Movecraft/src/main/java/net/countercraft/movecraft/Movecraft.java
@@ -23,7 +23,6 @@ import net.countercraft.movecraft.commands.*;
 import net.countercraft.movecraft.config.Settings;
 import net.countercraft.movecraft.craft.ChunkManager;
 import net.countercraft.movecraft.craft.CraftManager;
-import net.countercraft.movecraft.craft.datatag.CraftDataTagRegistry;
 import net.countercraft.movecraft.features.contacts.ContactsCommand;
 import net.countercraft.movecraft.features.contacts.ContactsManager;
 import net.countercraft.movecraft.features.contacts.ContactsSign;
@@ -58,7 +57,6 @@ public class Movecraft extends JavaPlugin {
     private WorldHandler worldHandler;
     private SmoothTeleport smoothTeleport;
     private AsyncManager asyncManager;
-    private  AbstractSignListener abstractSignListener;
     private WreckManager wreckManager;
 
     public static synchronized Movecraft getInstance() {
@@ -132,12 +130,6 @@ public class Movecraft extends JavaPlugin {
                     smoothTeleport = new BukkitTeleport(); // Fall back to bukkit teleportation
                     getLogger().warning("Falling back to bukkit teleportation provider.");
                 }
-            }
-
-            final Class<?> signListenerClass = Class.forName("net.countercraft.movecraft.compat." + WorldHandler.getPackageName(minecraftVersion) + ".SignListener");
-            if (AbstractSignListener.class.isAssignableFrom(signListenerClass)) {
-                abstractSignListener = (AbstractSignListener) signListenerClass.getConstructor().newInstance();
-                getServer().getPluginManager().registerEvents(abstractSignListener, this);
             }
         }
         catch (final Exception e) {
@@ -254,6 +246,7 @@ public class Movecraft extends JavaPlugin {
         AbstractMovecraftSign.register("Scuttle", new ScuttleSign());
         getServer().getPluginManager().registerEvents(new CraftPilotListener(), this);
         getServer().getPluginManager().registerEvents(new CraftReleaseListener(), this);
+        getServer().getPluginManager().registerEvents(new SignListener(), this);
         // Moved to compat section!
         //getServer().getPluginManager().registerEvents(new SignListener(), this);
 
@@ -367,7 +360,7 @@ public class Movecraft extends JavaPlugin {
         return asyncManager;
     }
 
-    public AbstractSignListener getAbstractSignListener() {return abstractSignListener;}
+    public SignListener getAbstractSignListener() {return abstractSignListener;}
 
     public @NotNull WreckManager getWreckManager(){
         return wreckManager;
